Some cleanup on tests so they work on Windows.

+++ autotest/t999_test.py
@@ -5,6 +5,9 @@ import config
 
 def test_teardown():
 
+    if config.retain:
+        return
+
     if os.path.isdir(config.dir_release):
         print('Removing folder ' + config.dir_release)
         shutil.rmtree(config.dir_release)
